consolidate CrcTools
